Add eniius.Eniius class as wrapper
Moved nxs_mcstas_inst.py to eniius/nexus.py
Changed scripts to use Eniius wrapper
